test-requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
coverage>=3.6
discover
hacking>=0.9.2,<0.10
mock>=1.0
oslotest>=1.2.0  # Apache-2.0
testrepository>=0.0.18
testtools>=0.9.36
fixtures>=0.3.14
requests>=2.2.0,!=2.4.0

# Documentation build requirements
sphinx>=1.1.2,!=1.2.0,!=1.3b1,<1.3
oslosphinx>=2.2.0  # Apache-2.0
